// src/store/modules/players/mutation-types.ts
export enum MutationTypes {
  SET_USER_NAME = "SET_USER_NAME",
  SET_ALL_WINNERS = "SET_ALL_WINNERS",
  SET_TODAY_WINNERS = "SET_TODAY_WINNERS",
  SET_YESTERDAY_WINNERS = "SET_YESTERDAY_WINNERS",
  SET_PLAYER_STATUS = "SET_PLAYER_STATUS",
  ADD_TODAY_WINNER = "ADD_TODAY_WINNER",
  SET_GET_WINNERS_REQUEST_STATUS = "SET_GET_WINNERS_REQUEST_STATUS",
}